Fix Wait to be idempotent and safe for concurrent use.
This change also fixes a race condition where goroutines could enter the group
after Wait finished waiting for the WaitGroup, but before it had cancelled the
context to signal that no further tasks should be accepted.  Now, if that
occurs, we will wait for the cancellation to propagate and for the straggler
tasks to exit before committing to the wait.