Greatly simplify the API of the group package.
This change removes the context plumbing from the group; the caller can handle
that in the usual manner (e.g., via lexical scope).  The callback mechanism for
errors remains and has been extended with some helpers, and throttling is now
handled with simple functions instead of a separate package.  There is no more
interface type, just the concrete type.