Simplify the implementation of the Collector.
Instead of maintaining a separate goroutine to synchronize delivery of values,
rework the collector to use a plain sync.Mutex.

This:

- Greatly simplifies the code (with one exception, noted below).

- Eliminates the need for a separate goroutine to service values. Each task now
  handles its own service, mediated by the collector. That, in turn:

- Eliminates the need to Wait for the Collector: Once all the goroutines
  running tasks in the collector have exited, the state is fully settled.
  The Wait method is now a no-op, and is marked as deprecated.

In addition, add a new Report method, replacing Stream. Instead of a channel,
tasks using this method accepts a report function that sends values to the
collector.  The report function ensures control does not return to the task
until the reported value has been serviced, which allows tasks to ensure they
do not exit until all their values have been addressed.

The Stream method still works, but is deprecated. To preserve its interface,
each Stream call now spins up a new goroutine to service the values from its
task. This is wasteful, but easily replaced by switching to Report.

Co-Authored-By: David Anderson <dave@natulte.net>