Simplify the implementation of the Group.
Remove the separate goroutine collecting errors, and deliver them directly to
the error filter and the output field. Moreover simplify the setup and teardown
so that there is not so much coordinated state. Although performance was not a
primary consideration, benchmarking suggests this is actually faster than the
previous implementation, and uses less memory.

Also expand and clarify the documentation of the Wait method.

Co-Authored-By: David Anderson <dave@natulte.net>