Make Collector naming more ergonomic.
I've gone back and forth on the terminology, but am settling on:

  Run:    func() or func() T          -- was NoError
  Call:   func() (T, error)           -- was Task
  Report: func(report func(T)) error  -- unchanged

This use of "Run" aligns with the top-level Run function, as well as the Run
method of the Group, which take no error.

This use of "Call" aligns with the top-level "Call" function, which wants a
value and an error.

Keep the existing method names as aliases, but mark them deprecated.

Also rename "NewCollector" as "Collect", which besides being shorter is also
more evocative of what it's doing. The construction is not the important aspect
of the collector, but what it provides.