taskgroup: drop *Group from the signature of start functions
While it is mildly convenient in a few cases to chain construction of the group
with starting a goroutine, it has turned out not to be worthwhile in practice.
Apart from the tests (which are hereby updated not to do that anymore), a
search of GitHub suggests nobody uses this -- even in code I wrote.