taskgroup: add an OnError method to Group
This allows an error filter to be set after the group is constructed.
The original New constructor is widely used, so it has been left alone, apart
from being rewritten in terms of the new method.

Apart from the added method, no functional changes.  Update a few tests to
tickle the edges of it a bit.