throttle: remove now-unnecessary leftover synchronization
Previously, the enter/release methods were exported, allowing for the
possibility that the caller could release twice on the same acquire.  To defend
against this, there was an atomic flag cleared by the first release.

Now that these methods are unexported, it is no longer possible to misuse them,
and that hedge isn't required.